refactored the structure
