+++ src/meteors/attr/lime.py
@@ -17,8 +17,8 @@ with warnings.catch_warnings():
 
 from meteors import HSI
 from meteors.attr.lime_base import Lime as LimeBase
-from meteors.utils.models import ExplainableModel, InterpretableModel, SkLearnLasso
-from meteors.utils.utils import torch_dtype_to_python_dtype, change_dtype_of_list, expand_spectral_mask
+from meteors.models import ExplainableModel, InterpretableModel, SkLearnLasso
+from meteors.utils import torch_dtype_to_python_dtype, change_dtype_of_list, expand_spectral_mask
 from meteors.attr import Explainer
 from meteors.attr import HSISpatialAttributes, HSISpectralAttributes
 from meteors.attr.attributes import ensure_torch_tensor
