+++ src/meteors/exceptions.py
