--- src/meteors/exceptions/__init__.py
@@ -1,14 +1,0 @@
-from .exceptions import (
-    ShapeMismatchError,
-    BandSelectionError,
-    MaskCreationError,
-    HSIAttributesError,
-)
-
-
-__all__ = [
-    "ShapeMismatchError",
-    "BandSelectionError",
-    "MaskCreationError",
-    "HSIAttributesError",
-]
