--- src/meteors/utils/__init__.py
@@ -1,6 +1,0 @@
-from .utils import agg_segmentation_postprocessing
-
-
-__all__ = [
-    "agg_segmentation_postprocessing",
-]
